requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
requests==2.31.0
nltk==3.8.1
psycopg2-binary
